+++ net-reporter/src/component/reporter.rs
@@ -13,6 +13,7 @@ use crate::config::Config;
 
 use crate::continuous_aggregate::http_request_methods_dist::HttpRequestMethodsDistAggregate;
 use crate::continuous_aggregate::network_bandwidth_per_protocol::NetworkBandwidthPerProtocolAggregate;
+use crate::continuous_aggregate::total_http_requests::TotalHttpRequestsAggregate;
 use crate::continuous_aggregate::ContinuousAggregate;
 use crate::continuous_aggregate::bandwidth_per_endpoint::BandwidthPerEndpointAggregate;
 use crate::continuous_aggregate::network_bandwidth::NetworkBandwidthAggregate;
@@ -24,6 +25,7 @@ use crate::query::charts::http_request_methods_dist::request::requester::HttpReq
 use crate::query::charts::network_bandwidth::request::requester::NetworkBandwidthRequester;
 use crate::query::charts::network_bandwidth_per_protocol::request::requester::NetworkBandwidthPerProtocolRequester;
 use crate::query::charts::network_graph::request::requester::NetworkGraphRequester;
+use crate::query::charts::total_http_requests::request::requester::TotalHttpRequestsRequester;
 use crate::query::filters::network_overview::request::requester::NetworkOverviewFiltersRequester;
 use crate::query::manager::query_manager::QueryManager; 
 
@@ -67,6 +69,7 @@ impl Reporter {
             .add_chart_generator(NetworkBandwidthRequester::default().boxed())
             .add_chart_generator(NetworkGraphRequester::default().boxed())
             .add_chart_generator(NetworkOverviewFiltersRequester::default().boxed())
+            .add_chart_generator(TotalHttpRequestsRequester::default().boxed())
             .add_chart_generator(HttpRequestMethodsDistRequester::default().boxed())
             .build()
     }
@@ -154,20 +157,36 @@ impl Reporter {
                 log::debug!("couldn't create {} refresh policy: {}", NetworkBandwidthPerProtocolAggregate::get_name(), err);
             }
         }
-        match HttpRequestMethodsDistAggregate::create(con).await {
+        match TotalHttpRequestsAggregate::create(con).await {
             Ok(_) => {
-                log::info!("successfully created {}", NetworkBandwidthPerProtocolAggregate::get_name());
+                log::info!("successfully created {}", TotalHttpRequestsAggregate::get_name());
             },
             Err(err) => {
-                log::debug!("couldn't create {}: {}", NetworkBandwidthPerProtocolAggregate::get_name(), err);
+                log::debug!("couldn't create {}: {}", TotalHttpRequestsAggregate::get_name(), err);
+            }
+        }
+        match TotalHttpRequestsAggregate::add_refresh_policy(con, None, None, "1 minute").await {
+            Ok(_) => {
+                log::info!("successfully created {}", TotalHttpRequestsAggregate::get_name());
+            },
+            Err(err) => {
+                log::debug!("couldn't create {}: {}", TotalHttpRequestsAggregate::get_name(), err);
+            }
+        }
+        match HttpRequestMethodsDistAggregate::create(con).await {
+            Ok(_) => {
+                log::info!("successfully created {}", HttpRequestMethodsDistAggregate::get_name());
+            },
+            Err(err) => {
+                log::debug!("couldn't create {}: {}", HttpRequestMethodsDistAggregate::get_name(), err);
             }
         }
         match HttpRequestMethodsDistAggregate::add_refresh_policy(con, None, None, "1 minute").await {
             Ok(_) => {
-                log::info!("successfully created {} refresh policy", NetworkBandwidthPerProtocolAggregate::get_name());
+                log::info!("successfully created {} refresh policy", HttpRequestMethodsDistAggregate::get_name());
             },
             Err(err) => {
-                log::debug!("couldn't create {} refresh policy: {}", NetworkBandwidthPerProtocolAggregate::get_name(), err);
+                log::debug!("couldn't create {} refresh policy: {}", HttpRequestMethodsDistAggregate::get_name(), err);
             }
         }
     }
@@ -228,6 +247,7 @@ impl Reporter {
 
         let response = query_manager.as_ref().handle_request(&config, recieve_enveloped_request, connection_pool).await;
         if response.is_err() {
+            log::error!("Error: {:?}", response.err());
             todo!()
         }
         let response = response.unwrap();