+++ net-reporter/src/query/charts/mod.rs
@@ -3,3 +3,4 @@ pub mod http_request_methods_dist;
 pub mod network_bandwidth;
 pub mod network_bandwidth_per_protocol;
 pub mod network_graph;
+pub mod total_http_requests;